Issue 41 - Refactor GitHubServiceFactory as an EJB/ CDI bean
Removed the GitHubServiceLoader class and the META-INF/service file
Using a CDI producer to instanciate the GitHubService implementation
Added integration test with Arquillian (deploying the GitHubService
within a war file, along with all required libs). The integration test
class extends the POJO test class to avoid code duplication.

The IT test may be run with the following command line:

    mvn clean verify -Pit,wildfly-managed-arquillian \
      -DGITHUB_TOKEN=$GITHUB_TOKEN_CATAPULT -DGITHUB_USERNAME=$GITHUB_USERNAME

where GITHUB_TOKEN_CATAPULT and GITHUB_USERNAME are the user's token
for the Catapult application, and his/her GitHub username.

Fixes #41