Issue 125 - Embed the project template
Added an 'openshift-project-template.json' file that contains the
template to apply on the OpenShift project.
Since there's no need anymore to look-up a 'pipelinetemplate.json'
file in the GitHub repo, all references to this argument in the code
(method parameters, etc.) could be removed.
In CatapultIT, the name of the OpenShift project is overriden with
a unique name to avoid collision with existing projects.

Fixes #125